lab2/server/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
psutil==5.9.6
